+++ src/components/fields/LocationChooserField.js
@@ -18,14 +18,11 @@ export default class LocationChooserField extends React.Component {
 				uiSchema: PropTypes.object,
 				taxonField: PropTypes.string,
 				geometryField: PropTypes.string,
-				strategy: PropTypes.string,
-				mapDrawOptions: PropTypes.shape({
-					marker: PropTypes.bool,
-					polyline: PropTypes.bool,
-					rectangle: PropTypes.bool,
-					polygon: PropTypes.bool,
-					circle: PropTypes.bool,
-				})
+				strategy: PropTypes.oneOf(["unit", "lolife", "lolifeUnit"]),
+				mapOptions: PropTypes.object,
+				maxShapes: PropTypes.number,
+				preselectMarker: PropTypes.bool,
+				color: PropTypes.string
 			})
 		}).isRequired,
 		schema: PropTypes.shape({
@@ -298,7 +295,7 @@ class LocationButton extends React.Component {
 						: geometryRef && geometryRef.type
 							? [geometryRef]
 							: [];
-					that.props.onChange(updateFormDataWithJSONPointer(that.props, [...geometries, event.feature.geometry], geometryField));
+					that.props.onChange(updateFormDataWithJSONPointer(that.props, {type: "GeometryCollection", geometries: [...geometries, event.feature.geometry]}, geometryField));
 				} else {
 					that.props.onChange(updateFormDataWithJSONPointer(that.props, event.feature.geometry, geometryField));
 				}